Account page refactor.
