+++ frontend/app/dashboard/settings/layout.tsx
@@ -3,7 +3,7 @@ import { SidebarNav } from "./components/sidebar-nav";
 
 const sidebarNavItems = [
     {
-      title: "Profile",
+      title: "Account",
       href: "/dashboard/settings",
     },
     {
@@ -23,7 +23,7 @@ interface SettingsLayoutProps {
 export default function SettingsLayout({children}: SettingsLayoutProps) {
     return (
         <>
-            <div className="hidden space-y-6 p-10 pb-16 md:block">
+            <div className="w-full space-y-6 p-10 pb-16 md:block">
                 <div className="space-y-0.5">
                     <h2 className="text-2xl font-bold tracking-tight">Settings</h2>
                     <p className="text-muted-foreground">
